Made output spreadsheets have integer values in the places where ArcGIS was expecting that.

+++ common.py
@@ -17,9 +17,9 @@ def constant_fn(val):
 
 def friendly_round(n):
     if not isinstance(n, numbers.Number): return n
-    if n >= 1000: return np.round(n, -3)
-    if n >= 100:  return np.round(n, -2)
-    if n >= 10:   return np.round(n, -1)
+    if n >= 1000: return int(np.round(n, -3))
+    if n >= 100:  return int(np.round(n, -2))
+    if n >= 10:   return int(np.round(n, -1))
     return 0
